move wrapped mui theme provider t components
